gh-911 add compound queries and geo queries

+++ test/integration/run.sh
@@ -2,8 +2,11 @@
 
 set -e 
 
-docker-compose down --remove-orphans
-
-docker-compose up -d esvector
+if [[ $* == *--no-restart* ]]; then
+  echo "Found --no-restart flag, reusing running dependencies from previous run..."
+else
+  docker-compose down --remove-orphans
+  docker-compose up -d esvector
+fi
 
 go test -count 1 -tags=integrationTest ./adapters/repos/...
